api/v1/booking/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter

from . import views

router = DefaultRouter()
router.register(r'', views.BookingViewSet)

urlpatterns = [
    path('', include(router.urls))
]
